put tracer into lib
