+++ rust_backend_test/src/bin/tester.rs
@@ -16,35 +16,18 @@ use tracing::{info, trace};
 use db::models::*;
 use rulaub_backend::{
     db,
-    logging::{file::get_logging_dir_path, util::create_env_filter},
+    logging::{logger::start_tracer, util::create_env_filter},
 };
 
 fn main() {
-    let log_dir = get_logging_dir_path().unwrap();
-    let file_appender = tracing_appender::rolling::daily(log_dir, "log");
-    let (non_blocking, _guard) = tracing_appender::non_blocking(file_appender);
-    let tracer = tracing_subscriber::fmt()
-        .with_writer(non_blocking)
-        .pretty()
-        .with_ansi(false)
-        // https://github.com/tokio-rs/tracing/issues/1310
-        .fmt_fields(tracing_subscriber::fmt::format::DefaultFields::new())
-        .with_level(true)
-        .with_target(true)
-        .with_thread_names(true)
-        .with_thread_ids(true)
-        // .with_max_level(Level::TRACE)
-        .with_env_filter(create_env_filter("TRACE"))
-        .with_filter_reloading();
-    let handle = tracer.reload_handle();
-    tracer.init();
+    let (tracing_handle, _guard) = start_tracer();
 
     trace!("App started.");
     info!("Logging level: TRACE");
 
     rulaub_backend::config::loader::load_and_watch_config_file();
 
-    handle.reload(create_env_filter("TRACE")).unwrap();
+    tracing_handle.reload(create_env_filter("TRACE")).unwrap();
 
     // let id = create_new_user();
     // println!("{}", id.unwrap());
