+++ rust_backend_test/src/logging.rs
@@ -1,2 +1,3 @@
 pub mod file;
+pub mod logger;
 pub mod util;
